requirements.txt
flask
line-bot-sdk
google-generativeai
youtube-transcript-api
groq
yt-dlp
gunicorn